BLOGG-backup/requirements.txt
asgiref==3.8.1
boto3==1.38.25
botocore==1.38.25
crispy-bootstrap4==2024.10
Django==5.2
django-crispy-forms==2.4
django-sslserver==0.22
django-storages==1.14.6
gunicorn==23.0.0
jmespath==1.0.1
packaging==25.0
pillow==11.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
waitress==3.0.2
